fix(p1-mfa-totp): fix critical security issues and split duplicate POST handlers
**CRITICAL FIXES:**

1. Client-controlled TOTP secret — FIXED
   - Old: Secret extracted from client-provided QR URL
   - New: Secret stored server-side during generate, verified against stored value
   - Impact: Prevents attackers from using arbitrary secrets

2. Duplicate POST handlers — FIXED
   - Old: 4 POST functions in one file (only last one works in Next.js)
   - New: Split into separate route files for each endpoint
   - Files created:
     - /api/auth/totp/route.ts (GET status only)
     - /api/auth/totp/generate/route.ts (POST generate)
     - /api/auth/totp/verify/route.ts (POST verify)
     - /api/auth/totp/disable/route.ts (POST disable)
     - /api/auth/totp/recovery/route.ts (POST recovery)

3. Recovery codes never consumed — FIXED
   - Updated consumeRecoveryCode() in lib/totp.ts
   - Old: bcrypt re-hash comparison (never matched)
   - New: Compare against stored hash, remove matching code from array
   - Impact: Recovery codes are now single-use

4. Removed unrelated SOC2_COMPLIANCE_REVIEW.md bundled in PR

All endpoints properly handle:
- Auth checks (getCurrentUser)
- Input validation
- Password verification for sensitive operations
- Error messages without leaking info
