+++ apps/web/src/app/api/auth/totp/disable/route.ts
@@ -1,0 +1,58 @@
+/**
+ * POST /api/auth/totp/disable — Disable TOTP MFA
+ *
+ * Requires password re-confirmation for security.
+ */
+
+import { NextRequest, NextResponse } from 'next/server'
+import { prisma } from '@/lib/db'
+import { getCurrentUser } from '@/lib/auth'
+import { compare } from 'bcryptjs'
+
+export async function POST(req: NextRequest) {
+  const user = await getCurrentUser()
+  if (!user) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+  }
+
+  const body = await req.json().catch(() => ({}))
+  const { password } = body as { password?: string }
+
+  if (!password) {
+    return NextResponse.json({ error: 'Password required to disable MFA' }, { status: 400 })
+  }
+
+  // Verify password
+  const dbUser = await prisma.user.findUnique({
+    where: { id: user.id },
+    select: { passwordHash: true },
+  })
+
+  if (!dbUser?.passwordHash) {
+    return NextResponse.json(
+      { error: 'Cannot disable MFA on accounts without password' },
+      { status: 400 },
+    )
+  }
+
+  const valid = await compare(password, dbUser.passwordHash)
+  if (!valid) {
+    return NextResponse.json({ error: 'Incorrect password' }, { status: 401 })
+  }
+
+  // Disable MFA and clear all TOTP data
+  await prisma.user.update({
+    where: { id: user.id },
+    data: {
+      totpEnabled: false,
+      totpSecret: null,
+      totpRecoveryCodes: null,
+      totpEnabledAt: null,
+    },
+  })
+
+  return NextResponse.json({
+    ok: true,
+    message: 'MFA disabled',
+  })
+}
